feat: вынес сущности в отдельные файлы
