+++ src/hooks/use-login.ts
@@ -3,7 +3,9 @@ import { navigate } from "@reach/router";
 import { useContext, useEffect } from "react";
 import { LogedInActionType, LogedInUser } from "../providers/loged-in-user";
 import type { LoggedUser, User } from "../entities/user";
-import { Credentials, Email, Password } from "../entities/credentials";
+import { Credentials } from "../entities/credentials";
+import { Password } from "../entities/password";
+import { Email } from "../entities/email";
 
 export default function useLogin(credentials: Credentials | null): User | null {
   const { loginService } = useContext(Services);
@@ -13,13 +15,13 @@ export default function useLogin(credentials: Credentials | null): User | null {
     if (!credentials || !dispatch) {
       return;
     }
-    loginService.login(
-      Email.from(credentials.email),
-      Password.from(credentials.password)
-    )
-      .then((user: LoggedUser) => dispatch!({ type: LogedInActionType.LOG_IN, payload: user }))
+    loginService
+      .login(Email.from(credentials.email), Password.from(credentials.password))
+      .then((user: LoggedUser) =>
+        dispatch!({ type: LogedInActionType.LOG_IN, payload: user })
+      )
       .then(() => navigate("/"))
-      .catch(e => alert(e.message));
+      .catch((e) => alert(e.message));
   }, [credentials, dispatch, loginService]);
 
   return state.user;
