+++ src/services/login-service.ts
@@ -1,14 +1,12 @@
-import { Email, Password } from "../entities/credentials";
 import { LoggedUser } from "../entities/user";
 import UserService from "./user-service";
+import { Email } from "../entities/email";
+import { Password } from "../entities/password";
 
 export default class LoginService {
   constructor(private readonly userService: UserService) {}
 
-  public async login(
-    email: Email,
-    password: Password
-  ): Promise<LoggedUser> {
+  public async login(email: Email, password: Password): Promise<LoggedUser> {
     const users = await this.userService.getAllUsers();
 
     for (let u of users) {
